moved the ChatManager to a seperate file
